pyproject.toml
[project]
name = "historical-drift-analyzer"
version = "0.1.0"
description = "Modular RAG-LLM pipeline"
authors = [
    {name = "Katharina Heller",email = "katharinaheller1@gmx.de"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pymupdf (>=1.26.5,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (==6.0.2)",
    "lxml (>=6.0.2,<7.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "spacy (==3.8.7)",
    "nameparser (>=1.1.3,<2.0.0)",
    "regex (>=2025.10.23,<2026.0.0)",
    "usaddress (>=0.5.16,<0.6.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "unstructured[pdf] (>=0.18.15,<0.19.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "clean-text (>=0.6.0,<0.7.0)",
    "habanero (>=2.3.0,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "keybert (>=0.9.0,<0.10.0)",
    "semanticscholar (>=0.11.0,<0.12.0)",
    "litstudy (>=1.0.6,<2.0.0)",
    "ragas (>=0.3.8,<0.4.0)",
    "datasets (>=4.4.1,<5.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "reportlab (>=4.4.4,<5.0.0)",
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "pymupdf (>=1.26.5,<2.0.0)"
]
ocr = [
    "ocrmypdf (>=15.0.0)"
]

